打包器简单完成
能够给通过源路径读取文件，最后将version放进目标文件夹，但是版本号还不能读取（dom->model的绑定有待实现）

+++ framework/lib/utils/XUtilsMath.cpp
@@ -4,7 +4,7 @@
 
 int XUtilsMath::getRand(int randMin, int randMax)
 {
-	int u = (double)rand() / (RAND_MAX + 1) * (randMax - randMin) + randMin;
+	int u = (int)rand() / (RAND_MAX + 1) * (randMax - randMin) + randMin;
 	return u;
 }
